defer to parent key to extract value when overriding, and coerce better
when we override we change e.g. the default value or the description,
but we don't intend to change how it extracts.  without this if we overwrite a
StructuredConfigKey we lose its smarts for extracting.

also when we coerce on write we didn't treat overwritten StructuredConfigKeys the
same as StructuredConfigKeys.  also we had overlooked the case where a List containing a Future
is being passed to a ConfigKey<Set>.

this repairs all those cases.